utils/requirements.txt
dotenv==0.9.9
PyYAML==6.0.2
requests==2.32.3
PyGithub==2.6.1